third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-create-from-nested-frame.html

<!DOCTYPE html>
<title>Credential Manager: Call create() across browsing contexts.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/webauth/authenticator.mojom.js"></script>
<script src="resources/credential-helpers.js"></script>
<body>
<script>

if (document.location.hostname == "127.0.0.1")
document.location = "https://subdomain.example.test:8443/credentialmanager/credentialscontainer-create-from-nested-frame.html";

function EncloseInScriptTag(str) {
  return "<script>" + str + "</scr" + "ipt>";
}

var TEST_NESTED_CREDENTIAL_ID = "nestedCredentialId";

// Use an invalid algorithm in the parameters for "success" cases
// so each test will exercise the rpID checks in  both the renderer
// and browser but return prior to reaching the device layer.
var CUSTOM_PUBLIC_KEY = 'var custom_public_key = '
    + '{challenge : new TextEncoder().encode("challenge"), '
    + 'rp: {id: "subdomain.example.test", name: "Acme"}, '
    + 'user: {id: new TextEncoder().encode("1098237235409872"), '
    + 'name: "acme@example.com", displayName: "Acme", icon:"iconUrl"}, '
    + 'pubKeyCredParams: [{type: "public-key", alg: 0,},], excludeCredentials:[],};';

var CREATE_CUSTOM_CREDENTIALS = CUSTOM_PUBLIC_KEY
    + "navigator.credentials.create({publicKey : custom_public_key})"
    + ".then(c => window.parent.postMessage(c.id, '*'))"
    + ".catch(e => window.parent.postMessage(e.name, '*'));";

var CREATE_CREDENTIALS = "navigator.credentials.create({publicKey})"
    + ".then(c => window.parent.postMessage(c.id, '*'));";
    
promise_test(t => {
  var PROBE_CREDENTIALS = "window.parent.postMessage(String(navigator.credentials), '*');";
  
  let frame = document.createElement("iframe");
  frame.src = "data:text/html," + EncloseInScriptTag(PROBE_CREDENTIALS);
  window.setTimeout(_ => document.body.append(frame));

  let eventWatcher = new EventWatcher(t, window, "message");
  return eventWatcher.wait_for("message").then(message => {
    assert_equals(message.data, "undefined");
  });
}, "navigator.credentials should be undefined in documents generated from `data:` URLs.");

promise_test(t => {
  let frame = document.createElement("iframe");
  frame.src = "resources/nested-mock-authenticator-client.html";
  window.setTimeout(_ => document.body.append(frame));

  let loadWatcher = new EventWatcher(t, frame, "load");
  loadWatcher.wait_for("load").then(_ =>
      frame.contentWindow.location = "javascript:" + CREATE_CREDENTIALS);

  let messageWatcher = new EventWatcher(t, window, "message");
  return messageWatcher.wait_for("message").then(message => {
    assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
  });
}, "navigator.credentials.create({publicKey}) in a javascript url should should succeed.");

var templateFrame = document.createElement("iframe");
templateFrame.src = "resources/nested-mock-authenticator-client.html";
templateFrame.addEventListener("load", _ => {

  promise_test(t => {
    let frame = document.createElement("iframe");
    frame.srcdoc = templateFrame.contentDocument.documentElement.outerHTML;
    window.setTimeout(_ => document.body.append(frame));

    let loadWatcher = new EventWatcher(t, frame, "load");
    loadWatcher.wait_for("load").then(_ => {
      frame.contentWindow.eval(CREATE_CREDENTIALS);
    });

    let eventWatcher = new EventWatcher(t, window, "message");
    return eventWatcher.wait_for("message").then(message => {
      assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
    });
  }, "navigator.credentials.create({publicKey}) in srcdoc should succeed.");

  promise_test(t => {
    let frame = document.createElement("iframe");
    frame.src = "about:blank";
    window.setTimeout(_ => document.body.append(frame));

    let loadWatcher = new EventWatcher(t, frame, "load");
    loadWatcher.wait_for("load").then(_ => {
      frame.contentDocument.write(templateFrame.contentDocument.documentElement.outerHTML);
      frame.contentDocument.write(EncloseInScriptTag(CREATE_CUSTOM_CREDENTIALS));
    });

    let eventWatcher = new EventWatcher(t, window, "message");
    return eventWatcher.wait_for("message").then(message => {
      assert_equals(message.data, TEST_NESTED_CREDENTIAL_ID);
    });
  }, "navigator.credentials.create({publicKey}) in an iframe embedded in a secure context should succeed.");

  promise_test(t => {
    let frame = document.createElement("iframe");
    frame.src = "about:blank";
    window.setTimeout(_ => document.body.append(frame));

    let loadWatcher = new EventWatcher(t, frame, "load");
    loadWatcher.wait_for("load").then(_ => {
      frame.contentWindow.eval(CREATE_CUSTOM_CREDENTIALS);
    });

    let eventWatcher = new EventWatcher(t, window, "message");
    return eventWatcher.wait_for("message").then(message => {
      assert_equals(message.data, "NotSupportedError");
    });
  }, "navigator.credentials.create({publicKey}) in an about:blank page embedded in a secure context should bypass rpID checks.");
});
document.body.append(templateFrame);

</script>
